LayoutTests/platform/gtk/http/tests/security/contentSecurityPolicy/1.1/securitypolicyviolation-basics-expected.txt

Check that a SecurityPolicyViolationEvent handler exists, and that events can be created.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL typeof document.onsecuritypolicyviolation should be object. Was undefined.
PASS typeof SecurityPolicyViolationEvent is "function"
PASS typeof window.e is "object"
PASS window.e.documentURI is "documentURIValue"
PASS window.e.referrer is "referrerValue"
PASS window.e.blockedURI is "blockedURIValue"
PASS window.e.violatedDirective is "violatedDirectiveValue"
PASS window.e.effectiveDirective is "effectiveDirectiveValue"
PASS window.e.originalPolicy is "originalPolicyValue"
PASS window.e.sourceFile is "sourceFileValue"
PASS window.e.lineNumber is 1
PASS window.e.columnNumber is 1
PASS window.e.statusCode is 200
PASS successfullyParsed is true

TEST COMPLETE
